Separate shipping models and address and register with modeladmin
