+++ longclaw/checkout/api.py
@@ -7,9 +7,8 @@ from rest_framework.decorators import api_view, permission_classes
 from rest_framework import permissions, status
 from rest_framework.response import Response
 from longclaw.basket.utils import get_basket_items, destroy_basket
-from longclaw.orders.models import Order, OrderItem, Address
-from longclaw.checkout.models import ShippingCountry
-from longclaw.checkout import serializers
+from longclaw.orders.models import Order, OrderItem
+from longclaw.shipping.models import Address
 from longclaw.checkout.utils import PaymentError
 from longclaw import settings
 
@@ -95,7 +94,6 @@ def capture_payment(request):
         )
         order_item.save()
 
-    
     try:
         gateway.create_payment(request, float(total)+postage)
         # Once the order has been successfully taken, we can empty the basket
@@ -108,70 +106,3 @@ def capture_payment(request):
         response = Response(data={"message": err.message, "order_id": order.id},
                             status=status.HTTP_400_BAD_REQUEST)
     return response
-
-class InvalidShippingRate(Exception):
-    pass
-
-class InvalidShippingCountry(Exception):
-    pass
-
-def get_shipping_cost(country_code, option):
-    try:
-        obj = ShippingCountry.objects.get(country_code=country_code)
-        if option == 'standard':
-            return {"rate": obj.standard_rate,
-                    "description": obj.standard_rate_description,
-                    "carrier": obj.standard_rate_carrier}
-        elif option == 'premium':
-            return {"rate": obj.premium_rate,
-                    "description": obj.premium_rate_description,
-                    "carrier": obj.premium_rate_carrier}
-        else:
-            raise InvalidShippingRate
-
-    except ShippingCountry.DoesNotExist:
-        if settings.DEFAULT_SHIPPING_ENABLED:
-            return {"rate": settings.DEFAULT_SHIPPING_RATE,
-                    "description": "Standard shipping to rest of world",
-                    "carrier": settings.DEFAULT_SHIPPING_CARRIER}
-        else:
-            raise InvalidShippingCountry
-
-
-
-@api_view(['GET'])
-@permission_classes({permissions.AllowAny})
-def shipping_cost(request):
-    ''' Returns the shipping cost for a given country
-    If the shipping cost for the given country has not been set, it will
-    fallback to the default shipping cost if it has been enabled in the app
-    settings
-    '''
-    try:
-        code = request.query_params.get('country_code')
-    except AttributeError:
-        return Response(data={"message": "No country code supplied"},
-                        status=status.HTTP_400_BAD_REQUEST)
-
-    option = request.query_params.get('shipping_option', 'standard')
-    try:
-        data = get_shipping_cost(code, option)
-        response = Response(data=data, status=status.HTTP_200_OK)
-    except InvalidShippingRate:
-        response = Response(data={"message": "Shipping option {} is invalid".format(option)},
-                            status=status.HTTP_400_BAD_REQUEST)
-    except InvalidShippingCountry:
-        response = Response(data={"message": "Shipping to {} is not available".format(code)},
-                            status=status.HTTP_400_BAD_REQUEST)
-
-    return response
-
-
-@api_view(["GET"])
-@permission_classes([permissions.AllowAny])
-def get_shipping_countries(request):
-    ''' Get all shipping countries
-    '''
-    queryset = ShippingCountry.objects.all()
-    serializer = serializers.ShippingCountrySerializer(queryset, many=True)
-    return Response(data=serializer.data, status=status.HTTP_200_OK)
