+++ longclaw/shipping/urls.py
@@ -1,0 +1,11 @@
+from django.conf.urls import url
+from longclaw.shipping import api
+
+urlpatterns = [
+    url(r'shipping/cost/$',
+        api.shipping_cost,
+        name='shipping_cost'),
+    url(r'shipping/countries/$',
+        api.shipping_countries,
+        name="shipping_countries")
+]
